@@ -39,150 +39,150 @@
     "react-dom": ">=16.0.0"
   },
   "dependencies": {
-    "@ant-design/icons": "~1.1.5",
-    "@ant-design/icons-react": "~1.1.1",
-    "array-tree-filter": "^2.0.0",
+    "@ant-design/icons": "~1.1.15",
+    "@ant-design/icons-react": "~1.1.2",
+    "array-tree-filter": "^2.1.0",
     "babel-runtime": "6.x",
-    "classnames": "~2.2.0",
-    "create-react-class": "^15.6.0",
-    "create-react-context": "0.2.2",
-    "css-animation": "^1.2.5",
+    "classnames": "~2.2.6",
+    "create-react-class": "^15.6.3",
+    "create-react-context": "0.2.3",
+    "css-animation": "^1.4.1",
     "dom-closest": "^0.2.0",
-    "enquire.js": "^2.1.1",
+    "enquire.js": "^2.1.6",
     "intersperse": "^1.0.0",
-    "lodash": "^4.17.5",
-    "moment": "^2.19.3",
+    "lodash": "^4.17.11",
+    "moment": "^2.22.2",
     "omit.js": "^1.0.0",
-    "prop-types": "^15.5.7",
+    "prop-types": "^15.6.2",
     "raf": "^3.4.0",
-    "rc-animate": "^2.4.1",
+    "rc-animate": "^2.5.4",
     "rc-calendar": "~9.7.9",
     "rc-cascader": "~0.16.0",
     "rc-checkbox": "~2.1.5",
     "rc-collapse": "~1.10.0",
-    "rc-dialog": "~7.2.0",
-    "rc-drawer": "~1.7.3",
-    "rc-dropdown": "~2.2.0",
-    "rc-editor-mention": "^1.0.2",
-    "rc-form": "^2.1.0",
-    "rc-input-number": "~4.1.0",
-    "rc-menu": "~7.4.1",
+    "rc-dialog": "~7.2.1",
+    "rc-drawer": "~1.7.6",
+    "rc-dropdown": "~2.2.1",
+    "rc-editor-mention": "^1.1.7",
+    "rc-form": "^2.2.3",
+    "rc-input-number": "~4.2.0",
+    "rc-menu": "~7.4.12",
     "rc-notification": "~3.2.0",
     "rc-pagination": "~1.17.3",
-    "rc-progress": "~2.2.2",
-    "rc-rate": "~2.4.0",
-    "rc-select": "~8.3.0",
-    "rc-slider": "~8.6.0",
+    "rc-progress": "~2.2.6",
+    "rc-rate": "~2.4.2",
+    "rc-select": "~8.4.0",
+    "rc-slider": "~8.6.3",
     "rc-steps": "~3.3.0",
     "rc-switch": "~1.8.0",
-    "rc-table": "~6.3.2",
-    "rc-tabs": "~9.4.0",
+    "rc-table": "~6.3.4",
+    "rc-tabs": "~9.4.6",
     "rc-time-picker": "~3.4.0",
-    "rc-tooltip": "~3.7.0",
-    "rc-tree": "~1.14.5",
-    "rc-tree-select": "~2.3.0",
-    "rc-trigger": "^2.5.4",
+    "rc-tooltip": "~3.7.3",
+    "rc-tree": "~1.14.6",
+    "rc-tree-select": "~2.3.1",
+    "rc-trigger": "^2.6.2",
     "rc-upload": "~2.6.0",
-    "rc-util": "^4.0.4",
-    "react-lazy-load": "^3.0.12",
-    "react-lifecycles-compat": "^3.0.2",
+    "rc-util": "^4.5.1",
+    "react-lazy-load": "^3.0.13",
+    "react-lifecycles-compat": "^3.0.4",
     "react-slick": "~0.23.1",
-    "shallowequal": "^1.0.1",
-    "warning": "~4.0.1"
+    "shallowequal": "^1.1.0",
+    "warning": "~4.0.2"
   },
   "devDependencies": {
     "@types/classnames": "^2.2.6",
-    "@types/prop-types": "^15.5.4",
-    "@types/react": "^16.0.0",
-    "@types/react-dom": "^16.0.0",
-    "@types/react-intl": "^2.3.10",
+    "@types/prop-types": "^15.5.6",
+    "@types/react": "^16.4.15",
+    "@types/react-dom": "^16.0.8",
+    "@types/react-intl": "^2.3.11",
     "@types/react-slick": "^0.23.2",
     "@yesmeck/offline-plugin": "^5.0.5",
-    "ansi-styles": "^3.2.0",
-    "ant-design-palettes": "^1.0.0",
+    "ansi-styles": "^3.2.1",
+    "ant-design-palettes": "^1.1.3",
     "antd-theme-generator": "1.1.3",
-    "antd-tools": "^5.1.6",
-    "babel-cli": "^6.18.0",
-    "babel-eslint": "^9.0.0",
-    "babel-plugin-import": "^1.0.0",
+    "antd-tools": "^5.2.2",
+    "babel-cli": "^6.26.0",
+    "babel-eslint": "^10.0.1",
+    "babel-plugin-import": "^1.9.1",
     "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-0": "^6.16.0",
-    "bezier-easing": "^2.0.3",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "bezier-easing": "^2.1.0",
     "bisheng": "^0.28.0",
-    "bisheng-plugin-antd": "^0.16.0",
-    "bisheng-plugin-description": "^0.1.1",
-    "bisheng-plugin-react": "^0.6.0",
-    "bisheng-plugin-toc": "^0.4.0",
-    "commander": "^2.11.0",
-    "cross-env": "^5.0.3",
-    "css-split-webpack-plugin": "^0.2.3",
-    "dekko": "^0.2.0",
-    "delegate": "^3.1.2",
+    "bisheng-plugin-antd": "^0.16.5",
+    "bisheng-plugin-description": "^0.1.4",
+    "bisheng-plugin-react": "^0.6.3",
+    "bisheng-plugin-toc": "^0.4.4",
+    "commander": "^2.18.0",
+    "cross-env": "^5.2.0",
+    "css-split-webpack-plugin": "^0.2.6",
+    "dekko": "^0.2.1",
+    "delegate": "^3.2.0",
     "docsearch.js": "^2.5.2",
-    "dora-plugin-upload": "^0.3.1",
+    "dora-plugin-upload": "^0.3.2",
     "enquire-js": "^0.2.1",
-    "enzyme": "^3.1.0",
-    "enzyme-adapter-react-16": "^1.0.0",
-    "enzyme-to-json": "^3.1.2",
-    "eslint": "^5.1.0",
-    "eslint-config-airbnb": "^17.0.0",
-    "eslint-plugin-babel": "^5.0.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "6.1.1",
-    "eslint-plugin-markdown": "~1.0.0-beta.4",
-    "eslint-plugin-react": "^7.10.0",
+    "enzyme": "^3.7.0",
+    "enzyme-adapter-react-16": "^1.6.0",
+    "enzyme-to-json": "^3.3.4",
+    "eslint": "^5.6.1",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-plugin-babel": "^5.2.1",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
+    "eslint-plugin-markdown": "~1.0.0-beta.6",
+    "eslint-plugin-react": "^7.11.1",
     "eslint-tinker": "^0.5.0",
-    "fetch-jsonp": "^1.0.3",
-    "glob": "^7.1.1",
-    "immutability-helper": "^2.5.0",
+    "fetch-jsonp": "^1.1.3",
+    "glob": "^7.1.3",
+    "immutability-helper": "^2.8.1",
     "intersection-observer": "^0.5.0",
-    "jest": "^23.2.0",
-    "jsdom": "^12.0.0",
+    "jest": "^23.6.0",
+    "jsdom": "^12.1.0",
     "jsonml.js": "^0.1.0",
-    "lint-staged": "^7.0.0",
+    "lint-staged": "^7.3.0",
     "lz-string": "^1.4.4",
     "majo": "^0.6.2",
-    "mockdate": "^2.0.1",
-    "moment-timezone": "^0.5.5",
+    "mockdate": "^2.0.2",
+    "moment-timezone": "^0.5.21",
     "pre-commit": "^1.2.2",
-    "preact": "^8.2.5",
-    "preact-compat": "^3.17.0",
+    "preact": "^8.3.1",
+    "preact-compat": "^3.18.4",
     "querystring": "^0.2.0",
-    "rc-queue-anim": "^1.4.1",
-    "rc-scroll-anim": "^2.2.1",
-    "rc-tween-one": "^2.0.1",
-    "react": "^16.3.2",
-    "react-color": "^2.11.7",
-    "react-copy-to-clipboard": "^5.0.0",
+    "rc-queue-anim": "^1.6.6",
+    "rc-scroll-anim": "^2.5.5",
+    "rc-tween-one": "^2.2.10",
+    "react": "^16.5.2",
+    "react-color": "^2.14.1",
+    "react-copy-to-clipboard": "^5.0.1",
     "react-dnd": "^5.0.0",
     "react-dnd-html5-backend": "^5.0.1",
-    "react-document-title": "^2.0.1",
-    "react-dom": "^16.3.2",
-    "react-github-button": "^0.1.1",
-    "react-infinite-scroller": "^1.0.15",
-    "react-intl": "^2.0.1",
+    "react-document-title": "^2.0.3",
+    "react-dom": "^16.5.2",
+    "react-github-button": "^0.1.11",
+    "react-infinite-scroller": "^1.2.1",
+    "react-intl": "^2.7.0",
     "react-resizable": "^1.7.5",
-    "react-router-dom": "^4.2.2",
+    "react-router-dom": "^4.3.1",
     "react-sticky": "^6.0.3",
-    "react-sublime-video": "^0.2.0",
-    "react-virtualized": "~9.20.0",
-    "remark-frontmatter": "^1.1.0",
+    "react-sublime-video": "^0.2.5",
+    "react-virtualized": "~9.20.1",
+    "remark-frontmatter": "^1.3.0",
     "remark-parse": "^5.0.0",
     "remark-stringify": "^5.0.0",
-    "remark-yaml-config": "^4.0.1",
+    "remark-yaml-config": "^4.0.2",
     "reqwest": "^2.0.5",
-    "rimraf": "^2.5.4",
-    "scrollama": "^1.4.1",
-    "stylelint": "^9.5.0",
-    "stylelint-config-standard": "^18.0.0",
-    "typescript": "~3.0.1",
+    "rimraf": "^2.6.2",
+    "scrollama": "^1.4.4",
+    "stylelint": "^9.6.0",
+    "stylelint-config-standard": "^18.2.0",
+    "typescript": "~3.1.1",
     "unified": "^7.0.0",
     "values.js": "^1.0.3",
-    "webpackbar": "^2.6.1",
-    "xhr-mock": "^2.4.0",
-    "xhr2": "^0.1.3"
+    "webpackbar": "^2.6.3",
+    "xhr-mock": "^2.4.1",
+    "xhr2": "^0.1.4"
   },
   "scripts": {
     "test": "jest --config .jest.js",
