@@ -96,7 +96,7 @@ class Spin extends React.Component<SpinProps, SpinState> {
   componentDidMount() {
     const { spinning, delay } = this.props;
     if (shouldDelay(spinning, delay)) {
-      this.delayTimeout = window.setTimeout(() => this.setState({ spinning }), delay);
+      this.delayTimeout = window.setTimeout(this.delayUpdateSpinning, delay);
     }
   }
 
@@ -109,9 +109,12 @@ class Spin extends React.Component<SpinProps, SpinState> {
     }
   }
 
-  componentWillReceiveProps(nextProps: SpinProps) {
-    const currentSpinning = this.props.spinning;
-    const spinning = nextProps.spinning;
+  componentDidUpdate() {
+    const currentSpinning = this.state.spinning;
+    const spinning = this.props.spinning;
+    if (currentSpinning === spinning) {
+      return;
+    }
     const { delay } = this.props;
 
     if (this.debounceTimeout) {
@@ -123,16 +126,23 @@ class Spin extends React.Component<SpinProps, SpinState> {
         clearTimeout(this.delayTimeout);
       }
     } else {
-      if (spinning && delay && !isNaN(Number(delay))) {
+      if (shouldDelay(spinning, delay)) {
         if (this.delayTimeout) {
           clearTimeout(this.delayTimeout);
         }
-        this.delayTimeout = window.setTimeout(() => this.setState({ spinning }), delay);
+        this.delayTimeout = window.setTimeout(this.delayUpdateSpinning, delay);
       } else {
         this.setState({ spinning });
       }
     }
   }
+
+  delayUpdateSpinning = () => {
+    const { spinning } = this.props;
+    if (this.state.spinning !== spinning) {
+      this.setState({ spinning });
+    }
+  };
 
   render() {
     const { className, size, prefixCls, tip, wrapperClassName, ...restProps } = this.props;
