@@ -10,7 +10,7 @@ export default function getScroll(target: any, top: boolean): number {
   let ret = isWindow ? target[prop] : target[method];
   // ie6,7,8 standard mode
   if (isWindow && typeof ret !== 'number') {
-    ret = window.document.documentElement[method];
+    ret = window.document.documentElement![method];
   }
 
   return ret;
