@@ -219,7 +219,7 @@ export default class FilterMenu<T> extends React.Component<FilterMenuProps<T>, F
       [`${dropdownPrefixCls}-menu-without-submenu`]: !this.hasSubMenu(),
     });
     let { filterDropdown } = column;
-    if (filterDropdown && typeof filterDropdown === 'function') {
+    if (filterDropdown instanceof Function) {
       filterDropdown = filterDropdown({
         prefixCls: `${dropdownPrefixCls}-custom`,
         setSelectedKeys: (selectedKeys: Array<any>) => this.setSelectedKeys({ selectedKeys }),
