@@ -38,6 +38,7 @@ export default class DirectoryTree extends React.Component<DirectoryTreeProps, D
   };
 
   state: DirectoryTreeState;
+  tree: Tree;
   onDebounceExpand: (event: React.MouseEvent<HTMLElement>, node: AntTreeNode) => void;
 
   // Shift click usage
@@ -157,35 +158,23 @@ export default class DirectoryTree extends React.Component<DirectoryTreeProps, D
     this.setUncontrolledState(newState);
   }
 
+  setTreeRef = (node: Tree) => {
+    this.tree = node;
+  };
+
   expandFolderNode = (event: React.MouseEvent<HTMLElement>, node: AntTreeNode) => {
-    const { expandedKeys = [] } = this.state;
-    const { onExpand } = this.props;
-    const { eventKey = '', expanded, isLeaf } = node.props;
+    const { isLeaf } = node.props;
 
     if (isLeaf || event.shiftKey || event.metaKey || event.ctrlKey) {
       return;
     }
 
-    let newExpandedKeys: string[] = expandedKeys.slice();
-    const index = newExpandedKeys.indexOf(eventKey);
+    // Get internal rc-tree
+    const internalTree = this.tree.tree;
 
-    if (expanded && index >= 0) {
-      newExpandedKeys.splice(index, 1);
-    } else if (!expanded && index === -1) {
-      newExpandedKeys.push(eventKey);
-    }
-
-    this.setUncontrolledState({
-      expandedKeys: newExpandedKeys,
-    });
-
-    if (onExpand) {
-      onExpand(newExpandedKeys, {
-        expanded: !expanded,
-        node,
-        nativeEvent: event.nativeEvent,
-      });
-    }
+    // Call internal rc-tree expand function
+    // https://github.com/ant-design/ant-design/issues/12567
+    internalTree.onNodeExpand(event, node);
   }
 
   setUncontrolledState = (state: DirectoryTreeState) => {
@@ -204,6 +193,7 @@ export default class DirectoryTree extends React.Component<DirectoryTreeProps, D
     return (
       <Tree
         icon={getIcon}
+        ref={this.setTreeRef}
         {...props}
 
         prefixCls={prefixCls}
